fix: gosec errors

+++ transmitter/internal/config/config.go
@@ -19,7 +19,7 @@ type AppConfig struct {
 type AuthConfig struct {
 	SharedSecret           Secret   `mapstructure:"shared_secret"`
 	AllowedOrigins         []string `mapstructure:"allowed_origins"`
-	TokenExpirationMinutes uint     `mapstructure:"expiration_minutes"`
+	TokenExpirationMinutes int64    `mapstructure:"expiration_minutes"`
 }
 
 func Load() (*Config, error) {
